requirements.txt
captcha==0.3
Pillow==7.0.0
